refactor: reorganize test structure and rename ProviderRegistry test
- Move mock classes from tests/unit/Providers/ to tests/mocks/ directory
- Rename AiProviderRegistryTest to ProviderRegistryTest to match renamed class
- Update mock class namespaces to WordPress\AiClient\Tests\mocks
- Ensure tests/unit structure mirrors src/ structure properly
